Refactor Project component to be more reusable.
Refactored the component to accept the github and web addresses as parameter. Also make the icon for web not to display if the address is not pass in as a parameter. In addition, icon to be displayed in white color via '.social-project li a'. This makes the component reusable and with better user experience.